Tweaked Targetting + Range = X
Finally! Managed to tweak targetting so that it now only targets units
that come into ranged! Added a little buffer area, so that it actually
targets units that are slightly farther than current range, but this
allows the unit to never overpass the other unit.

Second change is that range is now fully dependant on the X Axis, which
means even if their Y is not near enough, the unit can damage other
units!